Store game matchup with check for existing matchup
This commit extracts a matchup map Ex. {:time ... :teams "tA-tB"} from
an incoming odds bundle. Performs a check for an existing matchup of the
same teams on a given date and returns the matchup_id for the game. This
id is used as a FK in the odds table.

Two issues with this approach are not being able to handle double headers
or games which cross the midinight boundry in local time. In the former case
you'll end up with two seperate games having the same matchup_id and in the
latter case you'll have the same game split across two matchup_ids. These
issues may be addressed in a later commit, or I may leave it swept under the
rug ;)