Add tests to check matchup storage behavior
This commit adds two tests which check the matchup row creation behavior. If
a matchup of the same teams is seen on the same day it's assumed to be the same
game and no row is created. If a matchup is seen for the same teams more than a
day appart a new matchup row is created.