Trigger odds alerts for home and away teams
This commit updates the storage code so each home/away line is checked against
any matching entries for a game in the alert registry. If a book is found with
a line that's over the threshold the send-alert function is called with the line
and the book. Send-alert will be configured to send slack messages in a future
commit.