Only register an alert when a user presses the regiser button
Prior to this commit there were alerts being registered multiple times while
a user was filling out the form in Slack. This happens because Slack sends a
POST req anytime a value is selected in a drop down. This commit now only
updates the alert registry when the register alert button is clicked. The
slack-register-alert handler could be updated to avoid doing any munging unless
the POST is from the register alert button. If pursued that work will happen in
a later commit.