Add tests for Slack game alert registration
This commit adds tests which check that Slack actions register alerts as
expected. There are two types of actions in a Slack message which will send
a POST req to the /slack-alert-sub endpoint: selecting a team via a drop down,
and pressing the "register-alert" button. These tests assert that only pressing
the "register-alert" button will add an element to the atom that's used to
track registered alerts throughout the day.