Rework how Slack action POST reqs are handled
This commit fixes an issue which could cause an intermediate 400 bad request
response to be sent to Slack when a user was half way through filling out a
game alert form. This was caused because validation against the alert-sub spec
was being performed on any action. If the action was picking the first team
from the drop down then the spec would fail because both teams are required for
the alert-sub to be valid. This commit adds a guard so validation is only
performed when the user hits the register-alert button. If any other type of
action is sent from Slack we ack the request and no nothing, waiting for the
user to fill out the rest of the form and hit register-alert button.