Add CSV export function
This commit adds an export-odds-csv function which pulls all the data in the
db and writes it to a sportsball-odds-data.csv file. All available odds data
for a game at a given time is included in one csv row. Each book name present
in sb-specs/book-keys is used to create a <book-name>-home/away entry in the
csv. Then the odds for each book are looked up in the odds row when creating
a row for the csv export. This may result in many null values in the a row
depending on the source of the odds data and availability of odds from a given
book at the time the data was collected.