Change Jetty :join? setting
Prior to this commit the :join? flag was incorrectly set to false. This meant
that the Jetty server wouldn't block the main thread once started. This
resulted in the finally block being hit causing the at-at thread pool to be
cleared out so the scheduled scraping task would only happen once. With this
change the default of true is taken for the :join? setting causing the server
to block the main thread. The finally block has also been removed as resetting
the at-at pool isn't needed when the program is shutting down.