Update core_test.clj to account for config changes
This commit changes how the test helpers and tests handle passing config
information. Most notably the dynamic vars which previously were used
store test confg for *db* and *app* have been removed. Now the config maps
are provided to various parts of the system via the call-with-test-config
and call-with-test-app-and-config test helpers.